requirements.txt
streamlit>=1.31.0
anthropic>=0.40.0
sqlalchemy>=2.0.0
slack-sdk>=3.27.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.2.0
hubspot-api-client>=9.0.0
requests>=2.31.0
pyyaml>=6.0
python-dateutil>=2.8.2
schedule>=1.2.0
